refactor(client): lib -> libs 네이밍 변경
